Concat css files instead of using angular-css
Angular-css module is applies css too slowly. The css does not kick in
before the new page is loaded, so there is a visible delay and disruption
as the page loads. Loading all css upfront solves this problem. The file
concatenated styles is not too large, so the performance hit in initial
page loading time is negligible